justification-service/requirements.txt
# FastAPI et serveur web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données et ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Validation et configuration
pydantic==2.5.0
pydantic-settings==2.0.3

# HTTP client pour intégration services
httpx==0.25.2
requests==2.31.0

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilitaires
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Gestion des fichiers
aiofiles==23.2.0
python-magic==0.4.27

# Email et notifications
emails==0.6.0
jinja2==3.1.2

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Développement
black==23.11.0
flake8==6.1.0
isort==5.12.0
